# api-gateway/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AeroLink Distributed Platform API",
    "version": "2.0.0",
    "description": "AeroLink Airline Systems Platform – Multi-Region Cloud API. Implements RESTful endpoints across 4 microservices: Booking (Aurora SQL), Baggage (DynamoDB NoSQL), Check-In, and Flight Operations. Secured with JWT/OAuth2 and RBAC.",
    "contact": { "name": "AeroLink Engineering", "email": "api@aerolink.io" },
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local Development Gateway" },
    { "url": "https://api.aerolink.io", "description": "AWS Production (US-East-1)" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "Flight": {
        "type": "object",
        "properties": {
          "flight_id": { "type": "string", "example": "AL-101" },
          "origin": { "type": "string", "example": "Colombo (CMB)" },
          "destination": { "type": "string", "example": "London (LHR)" },
          "available_seats": { "type": "integer", "example": 145 },
          "total_seats": { "type": "integer", "example": 150 },
          "departure_time": { "type": "string", "format": "date-time" },
          "arrival_time": { "type": "string", "format": "date-time" },
          "price": { "type": "number", "example": 450.00 },
          "status": { "type": "string", "enum": ["ON_TIME", "DELAYED", "BOARDING", "CANCELLED"] }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Authenticate – obtain JWT token",
        "description": "Issues a signed JWT valid for 2 hours. Role must be: Passenger, Staff, or Admin.",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username"],
                "properties": {
                  "username": { "type": "string", "example": "risindu" },
                  "role": { "type": "string", "enum": ["Passenger", "Staff", "Admin"], "example": "Admin" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "JWT token issued successfully" },
          "400": { "description": "Missing username" },
          "429": { "description": "Rate limit exceeded (20 attempts / 15 min)" }
        }
      }
    },
...
    "/api/events/stream": {
      "get": {
        "summary": "SSE real-time event stream",
        "description": "Server-Sent Events stream. Pass JWT via ?token= query param. Broadcasts SEAT_RESERVED, BAGGAGE_STATUS_UPDATED, PASSENGER_CHECKED_IN, FLIGHT_STATUS_UPDATED events.",
        "tags": ["System"],
        "parameters": [{ "name": "token", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "text/event-stream connection established" } }
      }
    },
    "/api/bookings/flights": {
      "get": {
        "summary": "List all available flights",
        "description": "Retrieves relational flight data from Aurora SQL simulation.",
        "tags": ["Booking Service"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Array of flights",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Flight" } } } } } }
          }
        }
      }
    },
    "/api/bookings/reserve": {
      "post": {
        "summary": "Reserve a flight seat (ACID transaction)",
        "description": "Atomically decrements seat count and creates booking record. Emits SEAT_RESERVED event on SSE stream.",
        "tags": ["Booking Service"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["flightId", "passengerName"],
                "properties": {
                  "flightId": { "type": "string", "example": "AL-101" },
                  "passengerName": { "type": "string", "example": "James Smith" },
                  "seatClass": { "type": "string", "enum": ["Economy", "Business", "First"], "example": "Economy" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Booking confirmed with assigned seat number" },
          "400": { "description": "No seats available or missing fields" },
          "404": { "description": "Flight not found" }
        }
      }
    },
...
    "/api/baggage/update": {
      "post": {
        "summary": "Update baggage status (Event-driven telemetry)",
        "description": "Performs a DynamoDB upsert. Emits BAGGAGE_STATUS_UPDATED to SSE stream.",
        "tags": ["Baggage Service"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["baggageId", "status"],
                "properties": {
                  "baggageId": { "type": "string", "example": "BAG-77102" },
                  "status": { "type": "string", "example": "LOADED_ON_AIRCRAFT" },
                  "location": { "type": "string", "example": "London (LHR)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Status synchronized" },
          "400": { "description": "Invalid status value" }
        }
      }
    },
...
    "/api/flights/schedule": {
      "get": {
        "summary": "Get full flight schedule",
        "tags": ["Flight Operations"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Complete flight schedule with gate assignments and aircraft details" } }
      }
    },
    "/api/flights/pricing": {
      "get": {
        "summary": "Dynamic pricing matrix",
        "description": "Returns real-time pricing based on seat occupancy using tiered multipliers (simulates Lambda pricing engine).",
        "tags": ["Flight Operations"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Dynamic price per flight with demand tiers" } }
      }
    },
    "/api/flights/update-status": {
      "post": {
        "summary": "Update flight status (Staff/Admin only)",
        "description": "Requires Staff or Admin role. Updates flight status and broadcasts FLIGHT_STATUS_UPDATED via SSE.",
        "tags": ["Flight Operations"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["flightId", "status"],
                "properties": {
                  "flightId": { "type": "string", "example": "AL-101" },
                  "status": { "type": "string", "enum": ["ON_TIME", "DELAYED", "BOARDING", "DEPARTED", "ARRIVED", "CANCELLED"] },
                  "gate": { "type": "string", "example": "B14" },
                  "delayMins": { "type": "integer", "example": 45 },
                  "delayReason": { "type": "string", "example": "Weather" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Status updated and broadcast" },
          "403": { "description": "Insufficient role (requires Staff or Admin)" }
        }
      }
    }
  }
}